% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genVentas.R
\name{genVentas}
\alias{genVentas}
\title{Generador de datos de ventas}
\usage{
genVentas(
  dataset_nombre = "Dataset01",
  negocio_tipo = "VARIOS",
  n = 120000,
  fecha_min = "2022/01/01",
  fecha_max = "2022/12/31",
  moneda = "USD"
)
}
\arguments{
\item{dataset_nombre}{String/texto, define el nombre de la carpeta que contendrá los datos generados. Predeterminado: "Dataset01".}

\item{negocio_tipo}{String/texto, define el tipo de negocio al cual se dedica la empresa simulada. Valores validos para este parámetro son: INFORMATICA, LIBROS, RESTAURANTE, TELECOMUNICACIONES, VESTIMENTA, VIDEOJUEGOS, VARIOS. Predeterminado: "VARIOS".}

\item{n}{Integer/número entero, determina la cantidad de transacciones totales deseadas. Predeterminado: 120000.}

\item{fecha_min}{String/texto, indica la fecha mínima a partir de la cual se generarán los datos. Debe indicarse en formato yyyy/mm/dd. Predeterminado: "2022/01/01".}

\item{fecha_max}{String/texto, indica la fecha máxima hasta la cual se generarán los datos. Debe indicarse en formato yyyy/mm/dd. Predeterminado: "2022/12/31".}

\item{moneda}{String/texto, señala la moneda en la cual estarán expresados los precios de los productos comercializados. Valores válidos para este parámetros son: USD, PYG. Predeterminado: "USD".}
}
\value{
1 carpeta definida por dataset_nombre exportada en el working directory. Dicha carpeta contiene 3 sub-carpetas: Tablas de dimensiones, tablas de ventas codificadas, tablas de ventas decodificadas.
}
\description{
genVentas simula la extracción de datos de ventas de una empresa. La función exporta los datos mensuales generados a una carpeta dentro del working directory, conteniendo:
    a. Tablas de dimensiones
    b. Tablas de ventas codificadas
    c. Tablas de ventas decodificadas con las tablas de dimensiones
}
\examples{
genVentas (dataset_nombre = "Dataset01", negocio_tipo = "VARIOS", n = 120000, fecha_min = "2022/01/01", fecha_max = "2022/12/31", moneda = "PYG")
genVentas ("miEmpresa", "2020/12/01", "2024/06/30", "VESTIMENTA")
}
